requirements.txt
anthropic>=0.18.0
pypdf2>=3.0.0
python-dateutil>=2.8.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
gspread>=5.7.0
